migrate most of model_config to common; remainder goes to model_config_utils
