+++ src/backend_manager.h
@@ -30,7 +30,7 @@
 #include <string>
 #include <unordered_map>
 #include "constants.h"
-#include "model_config.h"
+#include "triton/common/model_config.h"
 #include "server_message.h"
 #include "status.h"
 #include "tritonserver_apis.h"
