+++ src/ensemble_utils.h
@@ -29,7 +29,7 @@
 
 #include <deque>
 #include <unordered_map>
-#include "model_config.h"
+#include "triton/common/model_config.h"
 #include "model_config.pb.h"
 #include "model_repository_manager.h"
 #include "status.h"
