+++ src/sequence_batch_scheduler.h
@@ -35,7 +35,7 @@
 #include <unordered_map>
 #include "backend_model.h"
 #include "backend_model_instance.h"
-#include "model_config.h"
+#include "triton/common/model_config.h"
 #include "model_config.pb.h"
 #include "rate_limiter.h"
 #include "scheduler.h"
