+++ src/server.cc
@@ -39,7 +39,7 @@
 #include "constants.h"
 #include "cuda_utils.h"
 #include "model.h"
-#include "model_config.h"
+#include "triton/common/model_config.h"
 #include "model_config.pb.h"
 #include "model_config_utils.h"
 #include "model_repository_manager.h"
